DockerComposeSetup - Added cleanUp method
  * Added methods to stop and remove containers created by DockerComposeSetup
  * Added method to remove network created by DockerComposeSetup
