+++ src/test/resources/docker-compose.yml
@@ -3,12 +3,12 @@ services:
   container1:
     build: "dockerfiles/healthy"
     ports:
-      - "8081:80"
+      - "80"
   container2:
     build: "dockerfiles/healthy"
     ports:
-      - "8082:80"
+      - "80"
   container3:
     build: "dockerfiles/healthy"
     ports:
-      - "8083:80"
+      - "80"
